Issue exactly one screenview event for a page
Avoid using ReactComponent state to track completion, since we might not
always be dealing with a mounted component at the point that we want to
kick off the screenview event. We can check and modify a direct object
property atomically, and we don't use that property for rendering, so we
can remove the connection to component state.