Add the follow/unfollow functionality for contributors
I added a bit of extra data that will be useful to the
Account and Subreddit models, changed the loading behavior
so the top doesn't jump, and styled things as they
currently are.  Cosmetic changes will come in subsequent
work, but I at least wanted to get the functionality in.