Fix a problem where screenview events don't always fire
The getSubreddit() function was trying to get the
subreddit name from the first comment.  This fails
when we don't have a comment in the post we're looking
at.  As an alternative, we see if there's a postId
in the current urlParams and see if we have that
post in our state to get the subreddit name from.